0024186: Eliminate remaining compiler warnings in MSVC++ 2010 64 bit with warning level 4
Eliminated some warnings of type C4267 (conversion from Type1 to Type2: possible loss of data); size_t to int explicit casts added.
Fixed warning C4267 and 4244 in OSD* files

Change in OSD_FileNode and OSD_File reverted since different handle types are needed on Unix and Windows

+++ src/IGESData/IGESData_IGESReaderData.cxx
@@ -51,7 +51,7 @@ IGESData_IGESReaderData::IGESData_IGESReaderData
     void IGESData_IGESReaderData::AddGlobal
   (const Interface_ParamType atype, const Standard_CString aval)
 {
-  theparh->Append(aval,strlen(aval),atype,0);
+  theparh->Append(aval,(int)strlen(aval),atype,0);
 }
 
     void IGESData_IGESReaderData::SetGlobalSection ()
